Remove minipack/rails.rb
